\name{frequencies}
\alias{frequencies}
\docType{data}

\title{
Word frequencies (5000 most frequent words) of 401 novels.
}

\description{
Word frequencies (5000 most frequent words) of 401 novels in Dutch.
}

\usage{
data(frequencies)
}

\details{
.................................................
}

\source{
The dataset is a part of The Riddle of Literary Quality Project.
}


\author{
Karina van Dalen-Oskam, Joris van Zundert
}

\seealso{
\code{\link{get.columns}}, \code{\link{explain}}, \code{\link{books}}, 
\code{\link{reviews}}, \code{\link{respondents}}
}



\examples{
data(frequencies)

print(frequencies)
summary(frequencies)
}

\keyword{datasets}
