\name{respondents}
\alias{respondents}
\docType{data}

\title{
Respondents' Answers
}

\description{
The information about the reviewers that participated in the survey called The National Reader Survey (2013).
}

\usage{
data(respondents)
}

\details{
.................................................
}

\source{
The dataset is a part of The Riddle of Literary Quality Project.
}

\examples{
data(respondents)

print(respondents)
summary(respondents)
}
\keyword{datasets}
