\name{reviews}
\alias{reviews}
\docType{data}

\title{
Reviewers' scores
}

\description{
Reviewers' scores from the survey called The National Reader Survey (2013).
}

\usage{
data(reviews)
}

\details{
.................................................
}

\source{
The dataset is a part of The Riddle of Literary Quality Project.
}

\author{
Karina van Dalen-Oskam, Joris van Zundert
}

\seealso{
\code{\link{get.columns}}, \code{\link{explain}}, \code{\link{books}}, 
\code{\link{frequencies}}, \code{\link{respondents}}
}




\examples{
data(reviews)

print(reviews)
summary(reviews)

}

\keyword{datasets}
