\name{books}
\alias{books}
\docType{data}

\title{
Measurements of 401 novels.
}

\description{
Measurements (including word frequencies) of 401 novels in Dutch.
}

\usage{
data(books)
}

\details{
.................................................
}

\source{
The dataset is a part of The Riddle of Literary Quality Project.
}

\examples{
data(books)

print(books)
summary(books)
}

\keyword{datasets}
